backend/requirements.txt
# Core
supabase==2.3.4
redis==5.0.1
fastapi==0.109.0
uvicorn==0.27.0
websockets==12.0
python-dotenv==1.0.0

# Agent framework
langgraph==0.0.55
langchain-core==0.2.0

# ML
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
joblib==1.3.2

# Scheduling
apscheduler==3.10.4
